dividing component into different files
