+++ src/Board.js
@@ -1,0 +1,36 @@
+import React from 'react';
+import SquareExtern  from './Square';
+
+// import './index.css';
+
+const Board = (props) => {
+  // Check that, in react inspector, Square component is named as Square instead of SquareExtern
+  const renderSquare = (i) => {
+    return <SquareExtern
+            value={props.squares[i]}
+            onClick={() => props.onClick(i)}
+          />
+  }
+
+  return (
+    <div>
+      <div className="board-row">
+        {renderSquare(0)}
+        {renderSquare(1)}
+        {renderSquare(2)}
+      </div>
+      <div className="board-row">
+        {renderSquare(3)}
+        {renderSquare(4)}
+        {renderSquare(5)}
+      </div>
+      <div className="board-row">
+        {renderSquare(6)}
+        {renderSquare(7)}
+        {renderSquare(8)}
+      </div>
+    </div>
+  );
+}
+
+export { Board };
